[New] build: Add `make CONFIG=djgpp` support. [Ref] Add macro MPT_COMPILER_QUIRK_NO_WCHAR and make mptString.cpp use char32_t as an internal UCS4 character type if wchar_t is unsupported on the given platform. [Ref] Fix various places that assume int32 == signed int, algthough it can also be signed long int on 32bit platform, and in fact it is on DJGPP. [New] openmpt123: Add liballegro 4.2 audio backend for DJGPP. [New] build: Add `make dist-dos` target. [Mod] libopenmpt: Bump prerel version.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@10793 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Fastmix.cpp
@@ -252,7 +252,7 @@ struct MixLoopState
 			}
 		}
 
-		Limit(nSmpCount, 1u, nSamples);
+		Limit(nSmpCount, uint32(1u), nSamples);
 
 #ifdef MPT_BUILD_DEBUG
 		{
